Refactor some utilities
